finished auth

+++ src/main/java/org/oscar/gym/service/trainee/TraineeService.java
@@ -1,9 +1,11 @@
 package org.oscar.gym.service.trainee;
 
+import org.oscar.gym.dtos.LoginDTO;
 import org.oscar.gym.dtos.TraineeDTO;
 import org.oscar.gym.dtos.response.TraineeResponse;
 import org.oscar.gym.entity.Trainee;
 import org.oscar.gym.repository.trainee.TraineeRepository;
+import org.oscar.gym.security.IAuthenticator;
 import org.oscar.gym.utils.Mapper;
 import org.springframework.stereotype.Component;
 
@@ -12,17 +14,22 @@ public class TraineeService implements ITraineeService  {
 
     private final TraineeRepository repository;
     private final Mapper mapper;
+    private final IAuthenticator authenticator;
 
-    public TraineeService(TraineeRepository repository, Mapper mapper) {
+    public TraineeService(TraineeRepository repository, Mapper mapper, IAuthenticator authenticator) {
         this.repository = repository;
         this.mapper = mapper;
+        this.authenticator = authenticator;
     }
 
     public void saveTrainee(TraineeDTO dto){
         repository.saveEntity(dto);
     }
 
-    public TraineeResponse findTrainee(String username){
+    public TraineeResponse findTrainee(LoginDTO loginDTO, String username){
+        if(!authenticator.isAuthorized(loginDTO.getUsername(), loginDTO.getPassword())){
+            throw new UnsupportedOperationException("Sorry user not authorized");
+        }
         try {
             Trainee trainee = repository.findEntity(username);
             return mapper.mapTraineeResponse(trainee);
@@ -33,7 +40,10 @@ public class TraineeService implements ITraineeService  {
 
     }
 
-    public TraineeResponse updateTrainee(TraineeDTO traineeDTO, long id) {
+    public TraineeResponse updateTrainee(LoginDTO loginDTO,TraineeDTO traineeDTO, long id) {
+        if(!authenticator.isAuthorized(loginDTO.getUsername(), loginDTO.getPassword())){
+            throw new UnsupportedOperationException("Sorry user not authorized");
+        }
         try {
             Trainee trainee = repository.updateEntity(traineeDTO, id);
             return mapper.mapTraineeResponse(trainee);
@@ -43,7 +53,10 @@ public class TraineeService implements ITraineeService  {
         return null;
     }
 
-    public void deleteTrainee(String username){
+    public void deleteTrainee(LoginDTO loginDTO,String username){
+        if(!authenticator.isAuthorized(loginDTO.getUsername(), loginDTO.getPassword())){
+            throw new UnsupportedOperationException("Sorry user not authorized");
+        }
         repository.deleteEntity(username);
     }
